Fix #1886
There were multiple problems with previous approach of always
expecting explicit "sent" signal from async handler, otherwise
displaying the `/send` message as pending, the most serious
problem was that the command handler is only called once during
send time, and it's problematic to depend on the result of that
on the recepient side, which could be for example offline.

The safer approach is to always display `/send` command as
successful (normal display) and only show errors when explicitly
signalised through the async command handler.

This should fix the most urgent problem for now.

Another issue which is not yet solved is that command handler is
called only after preview was already requested, that should clearly
happen in opposite order, also message, should be commited for further
processing/persisting only after handler returns success code, either
synchronously, or asynchronously but as logic input/sent logic is very
entangled and complicated, I decided to hold off with that now and
rather implement it as a part of the #1710 PR.